Split out JSON and `Clock` utilities to avoid circular imports
